rename subpage files and update their imports
