+++ src/components/SubPage/GenomicCharacterization.tsx
